Document the new Mozilla certdata.txt url
There is no code to download it automatically but the mk-ca-bundle.pl
file document the url at the top.

The url documented is the one before mozilla moved to mercurial as VCS
and wasn't updated for 3 years. The new one is the certificate bundle
for the current stable release of Firefox.